chore(testio): move request function to Util
